% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG_list_helper.R
\name{cleanGList}
\alias{cleanGList}
\title{cleanGList}
\usage{
cleanGList(g, concat = "///", filter = NULL)
}
\arguments{
\item{g}{Gene vector}

\item{concat}{String used to concatenate multiple mapped genes. Default '///'}

\item{filter}{Genes to filter out. Default NULL}
}
\description{
Cleanup concatenated gene list to remove na, blanks, filter protein coding genes
Intended to use with microarray probe map from AnnotationDbi.
}
