% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lazier_functions.R
\name{countUni}
\alias{countUni}
\title{countUni}
\usage{
countUni(x)
}
\arguments{
\item{x}{input vector}
}
\value{
no of unique values in x
}
\description{
count unique values in a vector
}
\examples{
a = c(1,2,3,NA,6,NA,7,7,1)
countUni(a)
}
