% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_color_palettes.R
\name{custom_color_palette}
\alias{custom_color_palette}
\alias{show_custom_palettes}
\title{Custom Color Palettes}
\usage{
custom_color_palette(palette = "OceanFire")

show_custom_palettes(n = 11)
}
\arguments{
\item{palette}{Name of the color palette to use. Use "list" to see all available palettes or "aliases" to see palette aliases.}

\item{n}{Number of colors to display for each palette}
}
\value{
A function that generates a color ramp palette.
}
\description{
Custom color palettes for data visualization.
}
\section{Functions}{
\itemize{
\item \code{show_custom_palettes()}: Print all available palettes

}}
\examples{
colpal <- custom_color_palette("OceanFire")(11)
colpal
show_custom_palettes(11)
}
