% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_DEG_list.R
\name{DEGMatSumm}
\alias{DEGMatSumm}
\title{DEGMatSumm}
\usage{
DEGMatSumm(degMat, sort = FALSE)
}
\arguments{
\item{degMat}{degMat from convertDEGList2Matrix results}

\item{sort}{sort or no?}
}
\description{
DEG List to deg Count
}
