% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_analysis.R
\name{enricherForGeneListWrapper}
\alias{enricherForGeneListWrapper}
\title{enricherForGeneListWrapper}
\usage{
enricherForGeneListWrapper(
  glist,
  term,
  pcut = 0.1,
  qcut = 0.2,
  minGSSize = 5,
  maxGSSize = 100
)
}
\arguments{
\item{glist}{List of genes (Entrez or Symbol) with names upGene, dnGene, toGene}

\item{term}{Pathway gene set term in containing pathway list}

\item{pcut}{p-value cutoff. Feeds to pvalueCutoff.}

\item{qcut}{q-value cutoff (FDR adjusted). Feeds to qvalueCutoff.}

\item{minGSSize}{minimum gene set size for enricher}

\item{maxGSSize}{maximum gene set size for enricher}
}
\value{
List of enricher Object: up, down
}
\description{
Wrapper function of enricher for differentially expressed gene list
}
\examples{
\dontrun{
sapply(glist, length)
enricherForGeneListWrapper(glist, c2.kegg)
}
}
