% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tanimotoCoef.R
\name{enrichment}
\alias{enrichment}
\title{Enrichment factor}
\usage{
enrichment(A, B, total)
}
\arguments{
\item{A}{set input1 as vectors}

\item{B}{set input2 as vectors}

\item{total}{total set as vectors}
}
\description{
Calculates enrichment factor against random overlap number
}
\examples{
A = c(1:5)
B = c(3:7)
total = c(1:10)
enrichment(A, B)
}
