% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMAP_style_gsea.R
\name{enrichmentFactorForDataFrame}
\alias{enrichmentFactorForDataFrame}
\title{CMAP Style Enrichment Calculation}
\usage{
enrichmentFactorForDataFrame(
  data,
  enrich_element,
  rank_by,
  cutoff = 0.05,
  min_count = 3,
  psc = 1
)
}
\arguments{
\item{data}{input data frame}

\item{enrich_element}{category column from data to enrich}

\item{rank_by}{column used to sort rank}

\item{cutoff}{cut percentile. default 0.05}

\item{min_count}{minimum count to be valid. default 3}

\item{psc}{pseudocount when caclulating enrichment. default 1}
}
\value{
dataframe of named vectors for each category
}
\description{
Calculate enrichment factor from a dataframe by rank cutoff.
}
