% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_analysis.R
\name{enrLS2plotMat}
\alias{enrLS2plotMat}
\title{enrLS2plotMat}
\usage{
enrLS2plotMat(enrLS, ord.FUN = NULL)
}
\arguments{
\item{enrLS}{enricher object list}

\item{ord.FUN}{function for ordering output plotMat}
}
\description{
Formerly enrObjectTransform nset, fixed to no filter, ordering function
}
