% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_analysis2.R
\name{enrobj2Matrix}
\alias{enrobj2Matrix}
\title{Enrichment object list to matrix}
\usage{
enrobj2Matrix(enrobj, val.col = "pvalue", log = TRUE)
}
\arguments{
\item{enrobj}{Enrichment result list.}

\item{val.col}{Column to output.}

\item{log}{Log transform? (used if p value is outputted)}
}
\value{
List of enricher Object: up, down
}
\description{
Turn enrichment object list (generated by ClusterProfiler enricher or Gen_enrichment) into matrix of needed value for drawing heatmap.
}
