% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_analysis.R
\name{enrObjectTransform_1set}
\alias{enrObjectTransform_1set}
\alias{enrObjectTransform_nset}
\alias{enrHeatmapOnly_1set}
\alias{enrHeatmapOnly_nset}
\title{enrObjectTransform}
\usage{
enrObjectTransform_1set(enr_obj, pco = 0.1, max_row = 10)

enrObjectTransform_nset(enr_obj, pco = 0.1, max_row = 10)

enrHeatmapOnly_1set(plotMat, mtitle, clust = TRUE)

enrHeatmapOnly_nset(plotMat, max_row, mtitle, colpal = "up",
  clust = TRUE)
}
\arguments{
\item{enr_obj}{enricher object list for single set (names: up, down)}

\item{pco}{p-value cutoff Default .1}

\item{max_row}{maximum row of plotMat to return Default 10}

\item{plotMat}{Matrix for plotting from enrTransform functions}

\item{mtitle}{Main title for pathway analysis heatmaps}

\item{clust}{Cluster Heatmap or no?}

\item{colpal}{Choosing color palette for nset pathway analysis heatmap. ('up' or 'dn')}
}
\value{
Matrix to feed to enr_HeatmapOnly (plotMat)
}
\description{
enricher Object list format transformation for plotting
}
\section{Functions}{
\itemize{
\item \code{enrObjectTransform_nset}: For multiple dataset gene list

\item \code{enrHeatmapOnly_1set}: Heatmap draw from plotMat from enrObjectTransform_1set

\item \code{enrHeatmapOnly_nset}: Heatmap draw from plotMat from enrObjectTransform_nset
}}

\section{Drawing Heatmaps}{
NA
}

\examples{
\dontrun{
plotMat <- enrObjectTransform_1set(enr_obj)
enrHeatmapOnly_1set(plotMat)
}
}
