% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG_list_helper.R
\name{ent2sym}
\alias{ent2sym}
\title{ent2sym}
\usage{
ent2sym(genes)
}
\arguments{
\item{genes}{genes either in Entrez or Symbol}
}
\value{
genes
}
\description{
gene mapper for entrez and symbol
}
\examples{
ent2sym('TP53')
}
