% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG_list_helper.R
\name{extractGeneList}
\alias{extractGeneList}
\title{Gene list extraction from result}
\usage{
extractGeneList(resultsLS, fco, qco, cnt = NULL, remove_ambi = FALSE)
}
\arguments{
\item{resultsLS}{list of result data frames. (limma results default)}

\item{fco}{fold change cut-offs (NOT LOG). Used to filter logFC column.}

\item{qco}{adjusted p value cut-offs. Used to filter adj.P.Val column.}

\item{cnt}{DEG count constrains (not used).}

\item{remove_ambi}{Remove genes both in up and down?}
}
\value{
Nested list of DEGs
}
\description{
Get DEG list from list of results generated by limma or DESeq2.
}
