% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_analysis2.R
\name{Gen_enrichment}
\alias{Gen_enrichment}
\title{Gene list pathway analysis by hypergeometric test}
\usage{
Gen_enrichment(glist, refgmt, tglist, ncore = 1)
}
\arguments{
\item{glist}{List of query genes. (DEG list)}

\item{refgmt}{Reference gene set list. (Pathway set in list form)}

\item{tglist}{List of gene space, same length as glist.}

\item{ncore}{Number of cores to use.}
}
\value{
List of enricher Object: up, down
}
\description{
Run pathway analysis (hypergeometric test) for gene list. Wrapper function for hypergeoTestForGeneset.
}
