% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneMap.R
\docType{data}
\name{geneAlias}
\alias{geneAlias}
\title{Gene symbol alias dataframe for Lazy2}
\format{A dataframe of columns 'hgnc_symbol', 'entrez'}
\source{
\href{ftp://ftp.ncbi.nlm.nih.gov/gene/DATA/GENE_INFO/Mammalia/Homo_sapiens.gene_info.gz}{NCBI human gene info}
}
\usage{
data(geneAlias)
}
\description{
Gene mapper dataframe containing Entrez, Symbol, Ensembl (optional)
used for ent2sym and more gene mapper functions.
Needs raw-data check for updates
}
\examples{
data(geneAlias)
}
\keyword{genes}
