% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG_list_helper.R
\name{geneAliasMap}
\alias{geneAliasMap}
\title{???}
\usage{
geneAliasMap(genes, return_ent = FALSE, verbose = FALSE)
}
\arguments{
\item{genes}{genes either in Entrez or Symbol (human)}

\item{return_ent}{genes either in Entrez or Symbol}

\item{verbose}{print mapped percentage}
}
\value{
genes
}
\description{
gene synonym mapper
}
\examples{
geneAliasMap('CHOP')
}
