% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG_list_helper.R
\name{geneCount}
\alias{geneCount}
\title{Count number of genes in geneList}
\usage{
geneCount(geneList)
}
\arguments{
\item{geneList}{Nested DEG list. See example for gene list structure.}
}
\description{
Lazy function for gene number for geneList
}
\examples{
set.seed(1234)
geneList <- list(
  up = list(A = sample(letters, 10), B = sample(letters, 5), C = sample(letters, 4)),
  dn = list(A = sample(letters, 6), B = sample(letters, 15), C = sample(letters, 7)),
  to = list(A = letters, B = letters, C = letters)
)
geneCount(geneList)
}
