% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG_list_helper.R
\name{geneListSetOverlap}
\alias{geneListSetOverlap}
\alias{geneListSetOverlap.parallel}
\alias{geneListDistMat}
\alias{geneListDistMat_HM}
\title{geneListSetOverlap}
\usage{
geneListSetOverlap(geneList)

geneListSetOverlap.parallel(geneList, ncore)

geneListDistMat(pairs, value.var)

geneListDistMat_HM(
  distMat,
  name,
  show_axis = TRUE,
  mtitle = NULL,
  log = TRUE,
  km = NULL
)
}
\arguments{
\item{geneList}{nested gene list of DEG}

\item{ncore}{number of cores to use}

\item{pairs}{pairs dataframe from geneListSetOverlap}

\item{value.var}{variable in pairs to use as value in matrix. Feeds to reshape2::acast}

\item{distMat}{distance matrix from geneListDistMat}

\item{name}{name to feed to ComplexHeatmap::Heatmap}

\item{show_axis}{show column and row names?}

\item{mtitle}{main title}

\item{log}{log transform or no?}

\item{km}{kmeans cluster feed for ComplexHeatmap::Heatmap}
}
\value{
dataframe of overlap measures
}
\description{
DEG List overlap
}
\section{Functions}{
\itemize{
\item \code{geneListSetOverlap.parallel}: For parallel processing

\item \code{geneListDistMat}: Covert pairs dataframe from geneListSetOverlap to matrix for distance

\item \code{geneListDistMat_HM}: Draw heatmap for distance matrix
}}

\examples{
\dontrun{
    pairs <- geneListSetOverlap(geneList)
    distMat <- geneListDistMat(pairs, 'ef_dn')
    geneListDistMat(distMat)
}
}
