% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG_list_helper.R
\name{getOverlapDF}
\alias{getOverlapDF}
\title{Get pairwise overlap significance for a list}
\usage{
getOverlapDF(gls, tgls, unique_combn = FALSE)
}
\arguments{
\item{gls}{Unnested list of genes}

\item{tgls}{List of gene space for each entry in gls}

\item{unique_combn}{Only return unique combination (no replicates) Default FALSE, will be set to TRUE in later versions}
}
\value{
dataframe
}
\description{
Calculate genelist set overlap (replaces pairsOverlap).
Metric : hypergeometric test p value, enrichment factor, tanimoto coef.
}
\examples{
\dontrun{
pairdf <- getOverlapDF(geneList$up, geneList$to)
pairm <- reshape2::acast(pairdf, ix1 ~ ix2, value.var = "ef")
}
}
