% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_analysis2.R
\name{glist_enrichment}
\alias{glist_enrichment}
\alias{Gen_enrichment}
\title{Gene list pathway analysis by hypergeometric test}
\usage{
glist_enrichment(glist, refgmt, tglist, minGeneSet = 10, ncore = 1, ef.psc = 0)

Gen_enrichment(glist, refgmt, tglist, minGeneSet = 10, ncore = 1, ef.psc = 0)
}
\arguments{
\item{glist}{List of query genes. (DEG list)}

\item{refgmt}{Reference gene set list. (Pathway set in list form)}

\item{tglist}{List of background gene space, same length as glist.}

\item{minGeneSet}{Minimum number of genes in gene set. Default 10.}

\item{ncore}{Number of cores to use.}

\item{ef.psc}{Pseudocount option for calculating enrichment factors. Default 0.}
}
\value{
List of enricher Object: up, down
}
\description{
Run pathway analysis (hypergeometric test) for gene list. Wrapper function for hypergeoTestForGeneset.
}
