% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tanimotoCoef.R
\name{harmean}
\alias{harmean}
\title{Harmonic Mean}
\usage{
harmean(v, ...)
}
\arguments{
\item{v}{numeric vector}

\item{...}{further arguments to be passed to or from methods}
}
\description{
Calculate harmonic mean. Harmonic means are often used to calculate means of p-values.
}
\examples{
v <- seq(0.01,.99, .01)
harmean(v)
}
