% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tanimotoCoef.R
\name{hypergeo_test}
\alias{hypergeo_test}
\alias{hypergeoTest}
\title{Hypergeometric test}
\usage{
hypergeo_test(query, reference, bgspace)

hypergeoTest(query, reference, bgspace)
}
\arguments{
\item{query}{query set        (balls drawn)}

\item{reference}{reference set    (white balls)}

\item{bgspace}{background space (balls in urn)}
}
\value{
Returns a dataframe of 4 columns. pVal is p value returned by phyper.
oddsRatio gives number of observation by number of expectation. int and bg refers to
the number of intersected items and background space respectively.
By nature of phyper, reference and query input should be interchangeable.
}
\description{
Run one-tailed hypergeometric test (fisher test) with set inputs.
'hypergeo_test' uses hypergeometric distribution function phyper with lower.tail = FALSE fixed.
This is commonly used for testing significance in overlap between sets.
}
