% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lazy2-package.R
\docType{package}
\name{Lazy2-package}
\alias{Lazy2}
\alias{Lazy2-package}
\title{Lazy2: Lazy Package}
\description{
Lazy package of custom functions for personal use in bioinformatic analyses.
}
\author{
\strong{Maintainer}: Yeogha Yoon \email{yeogha.yoon@mssm.edu} (\href{https://orcid.org/0000-0003-0955-4344}{ORCID})

}
\keyword{internal}
