% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LazygeneInfo.R
\docType{data}
\name{LazygeneInfo}
\alias{LazygeneInfo}
\title{Gene Mapping dataframe for Lazy2}
\format{A dataframe of columns 'hgnc_symbol', 'entrez'}
\source{
\href{https://www.gencodegenes.org/human/}{Gencode Metadata}
}
\usage{
data(LazygeneInfo)
}
\description{
Gene mapper dataframe containing Entrez, Symbol, Ensembl (optional)
used for ent2sym and more gene mapper functions.
Needs raw-data check for updates
}
\examples{
data(LazygeneInfo)
}
\keyword{genes}
