% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG_list_helper.R
\name{pairsOverlap}
\alias{pairsOverlap}
\title{pairsOverlap
Calculate overlap metric for pairs
This is an internal function used for geneListSetOverlap}
\usage{
pairsOverlap(v, geneList)
}
\arguments{
\item{v}{length 2 index vector from pair created by expand.grid}

\item{geneList}{nested gene list of DEG}
}
\value{
1-row dataframe
}
\description{
pairsOverlap
Calculate overlap metric for pairs
This is an internal function used for geneListSetOverlap
}
