% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGSEA.R
\name{plotEnrichment2}
\alias{plotEnrichment2}
\alias{theme_common_gsea}
\title{plotEnrichment2}
\usage{
plotEnrichment2(
  gset,
  stats,
  nes = NULL,
  qv = NULL,
  gseaParam = 1,
  mtitle = NULL,
  ytitle = "",
  ticksSize = 0.4,
  base_size = 7,
  line.col = "green",
  lwd = 2,
  ylims = NULL,
  draw = TRUE,
  statbar = FALSE,
  seed = 1234
)

theme_common_gsea(base_size = 7)
}
\arguments{
\item{gset}{gene set to calculate enrichment in character vector.}

\item{stats}{gene level statistics. This should be a named vector where names match gene ids in gset.}

\item{nes}{NES provided for annotation. Runs fgsea if not provided.}

\item{qv}{adjusted p value provided for annotation. Runs fgsea if not provided.}

\item{gseaParam}{gseaParam from plotEnrichment. default 1}

\item{mtitle}{main title of the plot.}

\item{ytitle}{y axis name}

\item{ticksSize}{ticks width for gene set bars. default 0.4}

\item{base_size}{font base size. default 7}

\item{line.col}{line colour for GSEA running score. default "green"}

\item{lwd}{line width for GSEA running score. default 2}

\item{ylims}{y-axis limits. If NULL (default), limits will be set automatically.}

\item{draw}{draw plot on device. Return ggplot object if set to FALSE. default TRUE}

\item{statbar}{show statistics tick at bottom of plot. default FALSE}

\item{seed}{random seed for reproducibility. default 1234}
}
\value{
enrichment plot
}
\description{
Draw a plot of running score for GSEA enrichment plot. 
Function has been taken and modified from plotEnrichment function in fgsea package.
}
\section{Functions}{
\itemize{
\item \code{theme_common_gsea()}: ggplot theme used in plotEnrichment2

}}
