% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG_list_helper.R
\name{removeAmbigDEGs}
\alias{removeAmbigDEGs}
\title{Remove ambiguous DEGs}
\usage{
removeAmbigDEGs(geneList)
}
\arguments{
\item{geneList}{gene list}
}
\value{

}
\description{
Remove DEGs if they're both in Down and Up.
}
\examples{
\dontrun{
geneList <- removeAmbigDEGs(geneList)
}
}
