% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lazier_functions.R
\name{removeNAs}
\alias{removeNAs}
\title{removeNAs}
\usage{
removeNAs(x)
}
\arguments{
\item{x}{input vector}
}
\value{
NA removed vectors
}
\description{
Remove NA values from a vector
}
\examples{
a <- c(1, 2, 3, NA, 6, NA, 7)
removeNAs(a)
}
