% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG_list_helper.R
\name{removeZeroVar}
\alias{removeZeroVar}
\title{removeZeroVar}
\usage{
removeZeroVar(m, idx1, idx2, verbose = TRUE)
}
\arguments{
\item{m}{Matrix (eg. Gene expression matrix)}

\item{idx1}{Column index for group A}

\item{idx2}{Column index for group B}

\item{verbose}{verbose}
}
\value{
Matrix with zero variance samples removed (hopefully)
}
\description{
Remove genes with zero variances before t-test matrix
}
