% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tanimotoCoef.R
\name{tanimotoCoef}
\alias{tanimotoCoef}
\title{Tanimoto coefficients}
\usage{
tanimotoCoef(A, B)
}
\arguments{
\item{A}{set input1 as vectors}

\item{B}{set input2 as vectors}
}
\description{
Calculates Tanimoto coefficients given two set vectors
}
\examples{
A = c(1:5)
B = c(3:7)
tanimotoCoef(A, B)
}
