% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGSEA.R
\name{theme_common_gsea}
\alias{theme_common_gsea}
\title{theme_common_gsea}
\usage{
theme_common_gsea(...)
}
\arguments{
\item{...}{ignored}
}
\description{
Retired. This function has been removed from the package.
See \code{workstation/plotGSEA.R} for the archived implementation.
}
