% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_transparent2.R
\name{theme_transparent2}
\alias{theme_transparent2}
\title{Transparent theme for ggplot}
\usage{
theme_transparent2(base_size = 12, x_angle = 0)
}
\arguments{
\item{base_size}{font base size. default 12}

\item{x_angle}{rotate x labels}
}
\description{
Modified transparent theme from ggpubr
}
\examples{
\dontrun{
theme_set(theme_transparent2())
}
}
