% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_transparent2.R
\name{theme_transparent3}
\alias{theme_transparent3}
\title{Transparent theme for ggplot}
\usage{
theme_transparent3(
  base_size = 11,
  base_family = "",
  base_line_size = base_size/22,
  base_rect_size = base_size/22,
  complete = TRUE
)
}
\arguments{
\item{base_size}{font base size. default 12}

\item{base_family}{base font family.}

\item{base_line_size}{base line size. default base_size / 22}

\item{base_rect_size}{base rectangle size. default base_size / 22}

\item{complete}{whether to return complete theme object. default TRUE}
}
\value{
ggplot theme
}
\description{
Theme for transparent ggplot figures. function adopts from theme_grey.
}
\examples{
\dontrun{
theme_set(theme_transparent3())
}

}
