% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG_list_helper.R
\name{TtestWithMat}
\alias{TtestWithMat}
\title{TtestWithMat}
\usage{
TtestWithMat(m, idx1, idx2, alternative = "two.sided", na.rm = TRUE)
}
\arguments{
\item{m}{Gene expression matrix}

\item{idx1}{index to feed to first vector in t.test}

\item{idx2}{index to feed to second vector in t.test}

\item{alternative}{alternative argument for t.test}
}
\value{
Dataframe with results
t_stat : T statistic
t_pVal : t-test p value
medDiff : median difference
menDiff : mean difference
}
\description{
Row-wise t-Test for gene expression matrix
}
\examples{
\dontrun{
TtestWithMat(M)
}
}
