% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG_list_helper.R
\name{convertDEGList2Matrix}
\alias{convertDEGList2Matrix}
\title{convertDEGList2Matrix}
\usage{
convertDEGList2Matrix(glist, toSpace = NULL)
}
\arguments{
\item{toSpace}{total gene space}
}
\description{
DEG List to deg Count
}
