% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_DEG_list.R
\name{convertDEGList2Matrix}
\alias{convertDEGList2Matrix}
\title{convertDEGList2Matrix}
\usage{
convertDEGList2Matrix(geneList, toSpace = NULL)
}
\arguments{
\item{geneList}{geneList of DEG}

\item{toSpace}{total gene space}
}
\description{
DEG List to deg Count
}
