% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG_list_helper.R
\name{drawMA}
\alias{drawMA}
\alias{drawVol}
\title{Draw MA, volcano plot}
\usage{
drawMA(resultDF, qco, fco, ttl_pre, ylim = NULL)

drawVol(resultDF, qco, fco, ttl_pre, xlim = NULL)
}
\arguments{
\item{resultDF}{result dataframe generated from limma or DEseq2. If DESeq2, column name should be 'adj.P.Val','logFC','AveExpr'}

\item{qco}{adjusted p value cut off}

\item{fco}{fold change cut off}

\item{ttl_pre}{main title prefix}

\item{ylim}{y-axis limit}

\item{xlim}{x-axis limit}
}
\value{
plot
}
\description{
Draw MA plot from resultDF generated from limma (data.frame format). DESeq2 results should have column names matching limma.
}
\section{Functions}{
\itemize{
\item \code{drawVol()}: Draw volcano plot

}}
\examples{
\dontrun{
drawMA(resultDF, qco=0.1, fco=2.0, ttl_pre='title')
}
}
