% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG_list_helper.R
\name{ent2sym}
\alias{ent2sym}
\title{gene mapper for entrez and symbol}
\usage{
ent2sym(genes, geneMap = NULL)
}
\arguments{
\item{genes}{genes either in Entrez or Symbol (human)}

\item{geneMap}{gene mapper data frame. Uses LazygeneInfo if null.}
}
\value{
genes in wanted IDs
}
\description{
Map Entrez ID to Hugo symbol and vice versa.
Uses LazygeneInfo if no gene map data frame is provided at geneMap.
}
\examples{
ent2sym('TP53')
}
