% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG_list_helper.R
\name{ent2sym}
\alias{ent2sym}
\title{Gene ID mapper for entrez and symbol}
\usage{
ent2sym(genes, geneMap = NULL)
}
\arguments{
\item{genes}{genes either in Entrez or Symbol (human)}

\item{geneMap}{gene mapper data frame. Uses LazygeneInfo if null.}
}
\value{
genes in wanted IDs
}
\description{
Map Entrez ID to HGNC symbol and vice versa.
Uses LazygeneInfo if no gene ID information is provided at geneMap.
geneMap, if provided, should be a data frame with at least two columns named 'hgnc_symbol' and 'entrez'.
ent2sym is meant for quick ID conversion and uses match, which does not handle many-to-one or many-to-many relationships.
}
\examples{
ent2sym('TP53')
}
