% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_DEG_list.R
\name{geneListSetOverlap}
\alias{geneListSetOverlap}
\alias{geneListSetOverlap.parallel}
\alias{geneListDistMat}
\alias{geneListDistMat_HM}
\title{geneListSetOverlap}
\usage{
geneListSetOverlap(geneList)

geneListSetOverlap.parallel(geneList, ncore)

geneListDistMat(pairs, value.var)

geneListDistMat_HM(distMat, name, mtitle = NULL, log = TRUE,
  km = NULL)
}
\arguments{
\item{geneList}{}
}
\value{
dataframe of overlap measures
}
\description{
DEG List overlap
}
\section{Functions}{
\itemize{
\item \code{geneListSetOverlap.parallel}: For parallel processing

\item \code{geneListDistMat}: Covert pairs dataframe from geneListSetOverlap to matrix for distance

\item \code{geneListDistMat_HM}: Draw heatmap for distance matrix
}}

\section{Section}{
NA

NA
}

