% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tanimotoCoef.R
\name{hypergeoTest}
\alias{hypergeoTest}
\title{hypergeoTest}
\usage{
hypergeoTest(query, reference, bgspace)
}
\arguments{
\item{query}{query set        (balls drawn)}

\item{reference}{reference set    (white balls)}

\item{bgspace}{background space (balls in urn)}
}
\description{
Hypergeometric test for set inputs (because I'm an idiot)
}
