% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMT_enrichments.R
\name{hypergeoTestForGeneset}
\alias{hypergeoTestForGeneset}
\title{Hypergeometric test for gmt style list}
\usage{
hypergeoTestForGeneset(query, refGMT, gspace)
}
\arguments{
\item{query}{gene set to query (ex. Differentially Expressed Genes)}

\item{refGMT}{list of reference gene set (ex. Pathways)}

\item{gspace}{gene space to query in}
}
\value{
data frame
pVal      : hypergeometric test p values from phyper
logP      : -log10(p value)
oddsRatio : odds ratio
tanco     : tanimoto coefficient (Jaccard index)
int       : intersected item count
bg        : reference item background count
}
\description{
Gene list hypergeometric test against gmt format list of gene set. Can be used for custom pathway analysis or CMAP style query.
}
\examples{
\dontrun{
gset = c('A','B','C')
glist = list(ID1 = LETTERS[1:10], ID2 = LETTERS[4:25])
hypergeoTestForGeneset(gset, glist, LETTERS)
}
}
