% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMT_enrichments.R
\name{hypergeoTestForGeneset}
\alias{hypergeoTestForGeneset}
\title{hypergeoTestForGeneset}
\usage{
hypergeoTestForGeneset(query, refGMT, gspace)
}
\arguments{
\item{query}{gene set to query (ex Differentially Expressed Genes)}

\item{refGMT}{reference gene set in list format}

\item{gspace}{gene space to query in}
}
\value{
data frame
}
\description{
gene hypergeometric test against gmt format gene set
}
\examples{
\dontrun{
hypergeoTestForGeneset(glist, c2.kegg)
}
}
