% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMT_enrichments.R
\name{hypergeoTestForGeneset}
\alias{hypergeoTestForGeneset}
\alias{hypergeoTestForGeneset2}
\title{Hypergeometric test for gmt style list}
\usage{
hypergeoTestForGeneset(query, refGMT, gspace, minGeneSet = 10, ef.psc = 0)

hypergeoTestForGeneset2(
  query,
  refGMT,
  gspace,
  minGeneSet = 10,
  ncore = 1,
  ef.psc = 0
)
}
\arguments{
\item{query}{gene set to query (eg. Differentially Expressed Genes)}

\item{refGMT}{list of reference gene set (eg. Pathways)}

\item{gspace}{background gene space. Should contain all genes in query.}

\item{minGeneSet}{minimum size of gene set. This is used to filter refGMT. Default 10}

\item{ef.psc}{pseudocount when calculating enrichment factor (oddsRatio). Default 0}
}
\value{
data frame
\describe{
	  \item{pVal}{: hypergeometric test p values from phyper}
	  \item{logP}{: -log10(p value)}
	  \item{oddsRatio}{: odds ratio}
	  \item{tanco}{: tanimoto coefficient (Jaccard index)}
	  \item{int}{: intersected item count}
	  \item{gsRatio}{: gene set ratio (selected genes in gene set / selected genes)}
	  \item{bgRatio}{: background ratio (total genes in gene set / total gene space)}
}
}
\description{
Gene list hypergeometric test against gmt format list of gene set. Can be used for custom pathway analysis or CMAP style query.
}
\section{Functions}{
\itemize{
\item \code{hypergeoTestForGeneset2}: Using multiprocessing
}}

\examples{
\dontrun{
gset = c('A','B','C')
glist = list(ID1 = LETTERS[1:10], ID2 = LETTERS[4:25])
hypergeoTestForGeneset(gset, glist, LETTERS)
}
}
