% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGSEA.R
\name{plotEnrichment2}
\alias{plotEnrichment2}
\alias{theme_common_gsea}
\title{plotEnrichment2}
\usage{
plotEnrichment2(
  gset,
  stats,
  nes,
  qv,
  gseaParam = 1,
  mtitle = NULL,
  ytitle = "",
  ticksSize = 0.4,
  base_size = 7,
  line.col = "green",
  lwd = 2,
  ylims = NULL,
  draw = TRUE,
  statbar = FALSE
)

theme_common_gsea(base_size = 7)
}
\arguments{
\item{gset}{gene set to calculate enrichment in character vector.}

\item{stats}{gene level statistics. This should be a named vector where names match gene ids in gset.}

\item{nes}{NES provided for annotation. Runs fgsea if not provided.}

\item{qv}{adjusted p value provided for annotation. Runs fgsea if not provided.}

\item{gseaParam}{gseaParam from plotEnrichment.}

\item{mtitle}{main title}

\item{ytitle}{y axis name}

\item{ticksSize}{ticks lwd for geneset bars.}

\item{base_size}{font base size. default 7}

\item{line.col}{line colour for GSEA running score.}

\item{lwd}{line width for GSEA running score.}

\item{draw}{draw plot. default TRUE}

\item{statbar}{show statistics tick at bottom of plot.}
}
\value{
enrichment plot
}
\description{
Plot running score for GSEA enrichment plot. Function modified from plotEnrichment.
}
\section{Functions}{
\itemize{
\item \code{theme_common_gsea}: Using multiprocessing
}}

