% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMT_enrichments.R
\name{readGMT}
\alias{readGMT}
\title{readGMT}
\usage{
readGMT(fn, skip = 2, as.df = TRUE)
}
\arguments{
\item{fn}{gmt file path}

\item{skip}{fields to skip. Usually 2.}

\item{as.df}{return as data frame?}
}
\value{
gene set dataframe of 2 column or list
}
\description{
gmt file reader
}
\examples{
\dontrun{
readGMT('file/path.gmt')
}
}
