% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tanimotoCoef.R
\name{tanimotoCoef}
\alias{tanimotoCoef}
\title{Tanimoto coefficients}
\usage{
tanimotoCoef(A, B, T = NULL)
}
\arguments{
\item{A}{set input1 as vectors}

\item{B}{set input2 as vectors}

\item{T}{background space (if given, A and B are intersected with T before calculating, causing the function to run slower)}
}
\description{
Calculates Tanimoto coefficients (Jaccard Index) given two set vectors.
}
\examples{
A <- c(1:5)
B <- c(3:7)
tanimotoCoef(A, B)
}
